moved client & server logic into same file
